+++ backend/spring-boot-3/src/test/java/back/springbootdeveloper/seungchan/controller/ApiTest.java
@@ -1,7 +1,7 @@
 package back.springbootdeveloper.seungchan.controller;
 
 import back.springbootdeveloper.seungchan.controller.config.AttendanceListFromJson;
-import back.springbootdeveloper.seungchan.controller.config.TestMakeObject;
+import back.springbootdeveloper.seungchan.controller.testutills.TestMakeObject;
 import back.springbootdeveloper.seungchan.dto.request.*;
 import back.springbootdeveloper.seungchan.entity.*;
 import back.springbootdeveloper.seungchan.repository.*;
