+++ backend/spring-boot-3/src/test/java/back/springbootdeveloper/seungchan/controller/config/jwt/TokenProviderTest.java
@@ -2,7 +2,7 @@ package back.springbootdeveloper.seungchan.controller.config.jwt;
 
 import back.springbootdeveloper.seungchan.config.jwt.JwtProperties;
 import back.springbootdeveloper.seungchan.config.jwt.TokenProvider;
-import back.springbootdeveloper.seungchan.controller.config.TestMakeObject;
+import back.springbootdeveloper.seungchan.controller.testutills.TestMakeObject;
 import back.springbootdeveloper.seungchan.entity.UserInfo;
 import back.springbootdeveloper.seungchan.repository.UserRepository;
 import back.springbootdeveloper.seungchan.repository.UserUtilRepository;
